requirements.txt
faster-whisper
sounddevice
numpy
keyboard
PyYAML
soundfile
jiwer